Changes in statement
Line 40 - If we do not pass empty dependency array then the useEffect hook only runs when the UI is rendered for the first time.

But if we do not pass empty dependency array then useEffect is called "Each time UI is rendered" and not only once. If the dependency array is "empty" then useEffect is called only once.